+++ ui/src/main/java/com/kotcrab/vis/ui/util/async/AsyncTaskProgressDialog.java
@@ -14,31 +14,38 @@
  * limitations under the License.
  */
 
-package com.kotcrab.vis.editor.ui.dialog;
+package com.kotcrab.vis.ui.util.async;
 
-import com.badlogic.gdx.Gdx;
-import com.kotcrab.vis.editor.util.async.AsyncTask;
-import com.kotcrab.vis.editor.util.async.AsyncTaskListener;
+import com.kotcrab.vis.ui.Locales.CommonText;
 import com.kotcrab.vis.ui.util.TableUtils;
+import com.kotcrab.vis.ui.util.async.AsyncTask.Status;
 import com.kotcrab.vis.ui.util.dialog.Dialogs;
 import com.kotcrab.vis.ui.widget.VisLabel;
 import com.kotcrab.vis.ui.widget.VisProgressBar;
 import com.kotcrab.vis.ui.widget.VisWindow;
 
 /**
- * Dialog displayed with {@link AsyncTask} when doing tasks on another thread
+ * Dialog used to display progress of {@link AsyncTask} as standard VisUI window. Shows progress bar and status
+ * of currently executed task.
  * @author Kotcrab
  */
 public class AsyncTaskProgressDialog extends VisWindow {
-	private AsyncTaskListener listener;
+	private AsyncTask task;
 
+	/**
+	 * Creates new dialog, note that task will be automatically started. Created dialog must be manually added to stage,
+	 * preferably with {@link VisWindow#fadeIn()} animation.
+	 * @param title title used as window title
+	 * @param task task to be executed
+	 */
 	public AsyncTaskProgressDialog (String title, AsyncTask task) {
 		super(title);
+		this.task = task;
 		setModal(true);
 
 		TableUtils.setSpacingDefaults(this);
 
-		final VisLabel statusLabel = new VisLabel("Please wait...");
+		final VisLabel statusLabel = new VisLabel(CommonText.PLEASE_WAIT.get());
 		final VisProgressBar progressBar = new VisProgressBar(0, 100, 1, false);
 
 		defaults().padLeft(6).padRight(6);
@@ -46,46 +53,43 @@ public class AsyncTaskProgressDialog extends VisWindow {
 		add(statusLabel).padTop(6).left().row();
 		add(progressBar).width(300).padTop(6).padBottom(6);
 
-		task.setListener(new AsyncTaskListener() {
+		task.addListener(new AsyncTaskListener() {
 			@Override
 			public void progressChanged (int newProgressPercent) {
-				Gdx.app.postRunnable(() -> progressBar.setValue(newProgressPercent));
-				if (listener != null) listener.progressChanged(newProgressPercent);
+				progressBar.setValue(newProgressPercent);
 			}
 
 			@Override
-			public void messageChanged (String newMsg) {
-				Gdx.app.postRunnable(() -> statusLabel.setText(newMsg));
-				if (listener != null) listener.messageChanged(newMsg);
+			public void messageChanged (String message) {
+				statusLabel.setText(message);
 			}
 
 			@Override
 			public void finished () {
 				fadeOut();
-				if (listener != null) listener.finished();
 			}
 
 			@Override
-			public void failed (String reason) {
-				failed(reason, null);
-				if (listener != null) listener.failed(reason);
-			}
-
-			@Override
-			public void failed (String reason, Exception ex) {
-				Dialogs.showErrorDialog(getStage(), reason == null ? "Unknown error occurred" : reason, ex);
-				if (listener != null) listener.failed(reason, ex);
+			public void failed (String message, Exception exception) {
+				Dialogs.showErrorDialog(getStage(), exception.getMessage() == null ? CommonText.UNKNOWN_ERROR_OCCURRED.get() : exception.getMessage(), exception);
 			}
 		});
 
 		pack();
 		centerWindow();
 
-		task.start();
+		task.execute();
 	}
 
-	public AsyncTaskProgressDialog setTaskListener (AsyncTaskListener listener) {
-		this.listener = listener;
-		return this;
+	public AsyncTask getTask () {
+		return task;
+	}
+
+	public void addListener (AsyncTaskListener listener) {
+		task.addListener(listener);
+	}
+
+	public Status getStatus () {
+		return task.getStatus();
 	}
 }