+++ ui/CHANGES.md
@@ -1,6 +1,9 @@
 #### Version: 1.2.5-SNAPSHOT (LibGDX 1.9.4)
+- **Added**: `AsyncTask` API and `AsyncTaskProgressDialog` (incubating - API or behaviour may change)
 - **Added**: `PopupMenu.removeEveryMenu(Stage)`
 - **Changed**: Selection of menu item is removed when mouse pointer leaves popup menu structure
+- **I18N Changes**:
+   - Added `Common` bundle
 
 #### Version: 1.2.4 (LibGDX 1.9.4)
 - **Added**: `ListSelection#setListener`, `#setProgrammaticChangeEvents` (with getters)